refactor to use `sigstore/sigstore` utilities
